Add files via upload
K-Means Clustering

There are many models for clustering out there. In this notebook, we will be presenting the model that is considered one of the simplest models amongst them. Despite its simplicity, the K-means is vastly used for clustering in many data science applications, especially useful if you need to quickly discover insights from unlabeled data. In this notebook, you will learn how to use k-Means for customer segmentation.

Some real-world applications of k-means:
• Customer segmentation
• Understand what the visitors of a website are trying to accomplish
• Pattern recognition
• Machine learning
• Data compression

In this notebook we practice k-means clustering with 2 examples:
• k-means on a random generated dataset
• Using k-means for customer segmentation

Table of contents¶
• k-Means on a randomly generated dataset
1. Setting up K-Means
2. Creating the Visual Plot

• Customer Segmentation with K-Means
1. Pre-processing
2. Modeling
3. Insights